refactor(core): extract conversation database operations
