+++ crates/openwave-core/src/db/ops/mod.rs
@@ -1,0 +1,1 @@
+pub(in crate::db) mod conversation;
